+++ recipes/configure.rb
@@ -1,10 +1,7 @@
 require 'securerandom'
 include_recipe 'database'
 include_recipe 'postgresql::ruby'
-
-unless(node[:repmgr][:repmgr_node_id])
-  include_recipe 'repmgr::node_id_generator'
-end
+include_recipe 'repmgr::dumb_repmgr_id'
 
 # create rep user and rep db
 
@@ -15,7 +12,16 @@ if(node[:repmgr][:replication][:role] == 'master')
     node.save # make sure the password gets saved!
   end
 else
-  master_node = search(:node, 'replication_role:master').first
+  pass_assign = resources(:bash => 'assign-postgres-password')
+  pass_assign.action :nothing
+
+  master_node = discovery_search(
+    'replication_role:master',
+    :environment_aware => node[:repmgr][:replication][:common_environment],
+    :minimum_response_time => false,
+    :raw_search => true,
+    :empty_ok => false
+  )
   if(master_node)
     pg_pass = master_node[:repmgr][:replication][:user_password]
   end
@@ -27,7 +33,7 @@ template File.join(node[:repmgr][:pg_home], '.pgpass') do
   variables(
     :password => pg_pass || node[:repmgr][:replication][:user_password]
   )
-  mode '0600'
+  mode 0600
 end
 
 key_bag = if(node[:repmgr][:data_bag][:encrypted])
@@ -41,28 +47,28 @@ key_bag = if(node[:repmgr][:data_bag][:encrypted])
           end
 
 directory File.join(node[:repmgr][:pg_home], '.ssh') do 
-  mode '0755'
+  mode 0755
   owner node[:repmgr][:system_user]
   group node[:repmgr][:system_user]
 end
 
 file File.join(node[:repmgr][:pg_home], '.ssh/authorized_keys') do
   content key_bag['public_key']
-  mode '0644'
+  mode 0644
   owner node[:repmgr][:system_user]
   group node[:repmgr][:system_user]
 end
 
 file File.join(node[:repmgr][:pg_home], '.ssh/id_rsa') do
   content key_bag['private_key']
-  mode '0600'
+  mode 0600
   owner node[:repmgr][:system_user]
   group node[:repmgr][:system_user]
 end
 
 template File.join(node[:repmgr][:pg_home], '.ssh/config') do
   source 'ssh_config.erb'
-  mode '0644'
+  mode 0644
   owner node[:repmgr][:system_user]
   group node[:repmgr][:system_user]
   variables( :hosts => node[:repmgr][:ssh_ignore_hosts] )
@@ -73,7 +79,7 @@ directory File.dirname(node[:repmgr][:config_file_path])
 
 template node[:repmgr][:config_file_path] do
   source 'repmgr.conf.erb'
-  mode '0644'
+  mode 0644
 end
 
 if(node[:repmgr][:replication][:role] == 'master')
@@ -124,8 +130,9 @@ else
   node.set[:postgresql][:config][:max_standby_streaming_delay] = node[:repmgr][:replication][:max_streaming_delay]
 
   if(master_node)
+    node.default[:repmgr][:addressing][:master] = master_node[:ipaddress]
     file '/var/lib/postgresql/.ssh/known_hosts' do
-      content %x{ssh-keyscan #{master_node[:ipaddress]}}
+      content %x{ssh-keyscan #{node[:repmgr][:addressing][:master]}}
     end
   end
 end