+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## v0.1.1
+* Update repmgr node id implementation
+* Allow customized addressing when node[:ipaddress] is not desired
+* Ensure Postgresql is dead before starting sync (fixes stale pid issue)
+* Allow repmgr to be installed via package
+
 ## v0.1.0
 * Initial release